feat: setting up new player profile
