+++ projects/bellumgens/src/app/player-section/sc-player/sc-player.component.html
@@ -1,0 +1,1 @@
+<p>sc-player works!</p>,
