+++ projects/bellumgens/src/app/player-section/sc-player/sc-player.component.ts
@@ -1,0 +1,10 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'app-sc-player',
+  imports: [],
+  templateUrl: './sc-player.component.html',
+  styleUrl: './sc-player.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ScPlayerComponent { }
